Revert "feat: add hooks for protected and unprotected routes"
This reverts commit e450c3250527e0124674568b3952b783f66a55ce.
