+++ src/app/auth/(withoutAuth)/forgot-password/page.tsx
@@ -7,24 +7,25 @@ import {
   PageSubheading,
 } from '@/components/layout/pageHeading';
 
-import RegisterUserForm from '../components/RegisterUserForm';
+import ForgotPasswordForm from '../../components/ForgotPasswordForm';
 
-export default function Register() {
+export default function ForgotPassword() {
   return (
     <>
       <PageHeader className='mt-8'>
-        <PageHeading heading='Join Monkeys' />
-        <PageSubheading subheading='Create your Monkeys account' />
+        <PageHeading heading='Forgot Password' />
+        <PageSubheading subheading='Enter your email to reset your password' />
       </PageHeader>
       <Container className='max-w-screen-sm my-8 px-3 flex flex-col gap-8'>
-        <RegisterUserForm />
+        <ForgotPasswordForm />
+
         <div className='mt-4 text-center'>
-          <span className='font-dm_sans'>Already have an account? </span>
+          <span className='font-dm_sans'>Go back to </span>
           <Link
             href='/auth/login'
             className='font-dm_sans hover:underline opacity-80 text-brand-orange'
           >
-            Login instead
+            Login
           </Link>
         </div>
       </Container>
