+++ src/app/auth/(withoutAuth)/login/page.tsx
@@ -9,7 +9,7 @@ import {
 } from '@/components/layout/pageHeading';
 import { Button } from '@/components/ui/button';
 
-import LoginForm from '../components/LoginForm';
+import LoginForm from '../../components/LoginForm';
 
 export default function LoginPage() {
   return (
