+++ src/app/auth/components/LoginForm.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react';
 
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
+import { useRouter, useSearchParams } from 'next/navigation';
 
 import { useSession } from '@/app/session-store-provider';
 import PasswordInput from '@/components/input/PasswordInput';
@@ -28,6 +28,7 @@ import { z } from 'zod';
 
 export default function LoginForm() {
   const router = useRouter();
+  const params = useSearchParams();
   const { update } = useSession();
 
   const [loading, setLoading] = useState(false);
@@ -43,6 +44,7 @@ export default function LoginForm() {
   const onSubmit = async (values: z.infer<typeof loginSchema>) => {
     setLoading(true);
 
+    const callbackURL = params.get('callbackURL');
     try {
       const loginResponse = await login(values);
       update({
@@ -50,13 +52,18 @@ export default function LoginForm() {
         data: { user: loginResponse },
       });
 
+      if (callbackURL) {
+        router.replace(callbackURL);
+        return;
+      }
+
       toast({
         variant: 'success',
         title: 'Login Successful',
         description: 'You have successfully logged in. Welcome back!',
       });
 
-      //router.replace('/feed');
+      router.replace('/feed');
     } catch (err) {
       console.log(err);
       toast({
