+++ src/app/auth/components/SearchParams.tsx
