+++ src/app/auth/reset-password/page.tsx
