+++ src/app/layout.tsx
@@ -12,7 +12,6 @@ import './globals.css';
 import { SessionStoreProvider } from './session-store-provider';
 import SWRProvider from './swr-provider';
 import { ThemeProviders } from './theme-provider';
-import ValidateSession from './validate-session';
 
 const arvo = Arvo({
   weight: ['400', '700'],
@@ -62,6 +61,8 @@ const RootLayout = async ({
 }: Readonly<{
   children: React.ReactNode;
 }>) => {
+  const authData = await validate();
+
   return (
     <html lang='en' suppressHydrationWarning>
       <head>
@@ -72,8 +73,7 @@ const RootLayout = async ({
       >
         <Toaster />
         <SWRProvider>
-          <SessionStoreProvider>
-            <ValidateSession />
+          <SessionStoreProvider value={authData}>
             <ThemeProviders>
               <TooltipProvider>
                 <Navbar />
